requirements.txt
boto==2.48.0
boto3==1.4.8
botocore==1.8.6
certifi==2017.11.5
chardet==3.0.4
dill==0.2.5
docutils==0.14
feedparser==5.2.1
future==0.16.0
httplib2==0.10.3
idna==2.6
jmespath==0.9.3
lockfile==0.12.2
logger==1.4
luigi==2.7.1
Pebble==4.3.6
PyFunctional==1.0.0
python-daemon==2.1.2
python-dateutil==2.6.1
request==0.0.22
requests==2.18.4
rope==0.10.7
s3transfer==0.1.12
six==1.10.0
tabulate==0.7.7
tornado==4.5.2
urllib3==1.22
